+++ Core/Utilities/Service/ServiceDomainProvider.cs
@@ -1,0 +1,18 @@
+namespace Unity.Cloud.Common
+{
+    /// <summary>
+    /// Enum of all supported service domain providers.
+    /// </summary>
+    public enum ServiceDomainProvider
+    {
+        /// <summary>
+        /// Azure provider.
+        /// </summary>
+        Azure,
+
+        /// <summary>
+        /// GCP provider.
+        /// </summary>
+        GCP,
+    }
+}