+++ Core/Utilities/Service/ServiceHost.cs
@@ -1,0 +1,18 @@
+namespace Unity.Cloud.Common
+{
+    /// <summary>
+    /// Contains the values for <see cref="ServiceEnvironment"/> and <see cref="ServiceDomainProvider"/>.
+    /// </summary>
+    public struct ServiceHost
+    {
+        /// <summary>
+        /// The value for the <see cref="ServiceEnvironment"/>.
+        /// </summary>
+        public string EnvironmentValue;
+
+        /// <summary>
+        /// The value for the <see cref="ServiceDomainProvider"/>.
+        /// </summary>
+        public string ProviderValue;
+    }
+}